requirements.txt
Babel==2.8.0
blinker==1.4
boto3==1.12.33
botocore==1.15.33
click==7.1.1
docutils==0.15.2
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
passlib==1.7.2
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.2.1
mysqlclient==1.3.12
